Update and rename NewClass.java to Swap_Two_List_Nodes
The file consists of code base of how to swap two linked list nodes.
This can be done by two ways:
1. Simply swapping the data value of the two nodes.
Adv. Very Simple and fast
2.Swapping the two nodes of the LinkedList
It is not necessary that the LinkedList nodes only consists of the data value, It can also contains some other information, so it is always advisable to swap the whole nodes by changing their pointers.